drs to uint32

+++ block/slvalidator.go
@@ -232,7 +232,7 @@ func (v *SettlementValidator) NextValidationHeight() uint64 {
 }
 
 // validateDRS compares the DRS version stored for the specific height, obtained from rollapp params.
-func (v *SettlementValidator) validateDRS(stateIndex uint64, height uint64, version uint64) error {
+func (v *SettlementValidator) validateDRS(stateIndex uint64, height uint64, version uint32) error {
 	drs, err := v.blockManager.Store.LoadDRSVersion(height)
 	if err != nil {
 		return err
